+++ src/main/java/edu/fullerton/csu/jmtran/projectx/dao/IMessageDAO.java
@@ -5,7 +5,7 @@ import edu.fullerton.csu.jmtran.projectx.model.Message;
 import java.util.List;
 
 public interface IMessageDAO {
-    public Message get(Long id);
+    public Message get(int id);
     public List<Message> list();
     public void save(Message message);
 }
