+++ src/main/java/edu/fullerton/csu/jmtran/projectx/model/MessageMapper.java
@@ -10,7 +10,7 @@ public final class MessageMapper implements RowMapper<Message> {
     public Message mapRow(ResultSet rs, int rowNum) throws SQLException {
         Message message = new Message();
 
-        message.setId(rs.getLong("id"));
+        message.setId(rs.getInt("id"));
         message.setSubject(rs.getString("subject"));
         message.setMessage(rs.getString("message"));
